Make Work.OL an instance attribute
Currently, the Work class is defined inside the get_work_helper_class function. This enables setting Work.OL as a class attribute but means that every call to the function recreates the class. Making the function return an instance of the class initialized from the ol_context will allow the class to be moved out of the function, improving efficiency.

To maintain compatibility with the previous instantiation logic, the logic is moved to Work().__call__, making Work instances callable to provide olid, identifiers and kwargs.